Completed Observable

Rocket.completed — Function

completed(::Type{T} = Any; scheduler::H = AsapScheduler()) where { T, H <: AbstractScheduler }

Creation operator for the CompletedObservable that emits no items to the Actor and immediately sends a complete notification on subscription.

Arguments

T: type of output data source, optional, Any is the default

scheduler: optional, scheduler-like object

Examples

using Rocket

source = completed(Int)
subscribe!(source, logger())
;

# output

[LogActor] Completed

See also: CompletedObservable, subscribe!, logger

Rocket.CompletedObservable — Type

CompletedObservable{D, H}(scheduler::H)

Observable that emits no items to the Actor and just sends a complete notification on subscription.

Constructor arguments

scheduler: Scheduler-like object

See also: Subscribable, completed