Faulted Observable

Rocket.faulted — Function

faulted(err; scheduler::H = AsapScheduler())            where { H <: AbstractScheduler }
faulted(::Type{T}, err; scheduler::H = AsapScheduler()) where { T, H <: AbstractScheduler }

Creation operator for the FaultedObservable that emits no items to the Actor and immediately sends an error notification.

Arguments

err: the particular Error to pass to the error notification.

T: type of output data source, optional, Any by default

scheduler: optional, scheduler-like object

Examples

using Rocket

source = faulted("Error!")
subscribe!(source, logger())
;

# output

[LogActor] Error: Error!

See also: FaultedObservable, subscribe!

Rocket.FaultedObservable — Type

FaultedObservable{D, H}(err, scheduler::H)

Observable that emits no items to the Actor and just sends an error notification on subscription.

Constructor arguments

err: error to emit on subscription

scheduler: scheduler-like object

See also: faulted